Add files via upload
This is a copy of the completion code. Doesn't follow the line quite right. All you need to do is change your robot.cpp with the lines in this text file. To change the map to the completion map in the sever folder you need to open up the config.text and change .core to .completion.